Add support for forge servers
Forge gets mad when you try to connect before the server is up yet. To
fix this, we listen for kicks that contain the message for sends to the
client telling them the server isn't ready yet. 

There might be a better way to do this, but it works...

Also bump to Java 8, because you can go die in a fire if your still on 7